✨(app-desk) integrate team endpoint
Integrate team endpoint with react-query.
It will be used to get the team on the team page.
